backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0.post1
python-multipart>=0.0.6
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
python-docx>=0.8.11
pdfplumber>=0.10.2
markdownify>=0.11.6
requests>=2.31.0